requirements.txt
django-allauth
django-environ
django-rest-auth
django-rest-swagger
djangorestframework
djangorestframework_jwt
